# docs/openapi.yaml
openapi: 3.0.3

info:
  title: Music Service API
  description: API для онлайн библиотеки песен
  version: 1.0.0

servers:
  - url: http://localhost:8080/api
    description: Local server

paths:  
  /songs:
    put:
      summary: Добавить новую песню
      requestBody:
        description: Данные песни
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddSongRequest'
      responses:
        '200':
          description: Песня успешно добавлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Song'
        '400':
          description: Пустое имя или название группы
        '500':
          description: Internal server error

    delete:
      summary: Удалить песню
      parameters:
        - name: songID
          in: query
          required: true
          schema:
            type: integer
          description: Идентификатор песни
      responses:
        '200':
          description: Песня успешно удалена
        '400':
          description: Неверный запрос
        '404':
          description: Песня не найдена

    get:
      summary: Получить список песен с фильтрацией
      description: Получение списка песен с возможностью фильтрации по различным полям.
      parameters:
        
        - name: group
          in: query
          schema:
            type: string
          description: Название группы
          example: Madonna

        - name: release_date
          in: query
          schema:
            type: string
            format: date
          description: Дата релиза песни
          example: 2008-08-08
        
        - name: lyrics
          in: query
          schema:
            type: string
          description: Поиск совпадений в тексте песни (string matching)
          example: incredible
        
        - name: link
          in: query
          schema:
            type: string
            format: uri
          description: Ссылка на песню
          example: https://example.com/incredible
        
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            minimum: 0
          description: Количество песен для отображения
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Смещение для пагинации
      responses:
        '200':
          description: Успешное получение списка песен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Song'
        '400':
          description: Неверный запрос
        '500':
          description: Внутренняя ошибка сервера


    patch:
      summary: Обновить песню
      requestBody:
        description: Данные для обновления песни
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSongRequest'
      responses:
        '200':
          description: Песня успешно обновлена
        '400':
          description: Неверный запрос
        '404':
          description: Песня не найдена

  /lyrics:
    get:
      summary: Получить текст песни с пагинацией
      parameters:
        - name: song_id
          in: query
          required: true
          schema:
            type: integer
          description: Идентификатор песни
          example: 7
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Начинать с куплета
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
          description: Количество куплетов
      responses:
        '200':
          description: Текст песни
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetLyricsResponse'
        '400':
          description: Неверный запрос

components:
  schemas:
    AddSongRequest:
      type: object
      required:
        - song
        - group
      properties:
        song:
          type: string
          example: Hey Jude
        group:
          type: string
          example: The Beatles

    UpdateSongRequest:
      type: object
      required:
        - song_id
      properties:
        song_id:
          type: integer
          example: 1
        new_name:
          type: string
          example: Hey Jude Updated
        release_date:
          type: string
          format: date
          example: 1968-08-30
        lyrics:
          type: string
          example: "Hey Jude, don't make it bad..."
        link:
          type: string
          example: https://example.com/hey-jude

    Song:
      type: object
      properties:
        ID:
          type: integer
        Name:
          type: string
        ReleaseDate:
          type: string
          format: date
        Lyrics:
          type: string
        Link:
          type: string
        Group:
          type: object
          properties:
            ID:
              type: integer
            Name:
              type: string

    GetLyricsResponse:
      type: object
      properties:
        verses:
          type: array
          items:
            type: string
          example:
            - "Hey Jude, don't make it bad."
            - "Take a sad song and make it better."